+++ src/main/java/de/codeinfection/quickwango/AntiGuest/Commands/ReloadCommand.java
@@ -2,6 +2,8 @@ package de.codeinfection.quickwango.AntiGuest.Commands;
 
 import de.codeinfection.quickwango.AntiGuest.AbstractCommand;
 import de.codeinfection.quickwango.AntiGuest.BaseCommand;
+import de.codeinfection.quickwango.AntiGuest.Prevention;
+import de.codeinfection.quickwango.AntiGuest.PreventionManager;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.Plugin;
@@ -22,18 +24,43 @@ public class ReloadCommand extends AbstractCommand
     @Override
     public String getDescription()
     {
-        return "Reloads the configuration of the plugin.";
+        return "Reloads the configuration of the plugin or a specific prevention.";
+    }
+
+    @Override
+    public String getUsage()
+    {
+        return super.getUsage() + " [prevention]";
     }
 
     @Override
     public boolean execute(CommandSender sender, String[] args)
     {
-        final Plugin plugin = getBase().getPlugin();
-        final PluginManager pm = plugin.getServer().getPluginManager();
-        pm.disablePlugin(plugin);
-        pm.enablePlugin(plugin);
+        if (args.length > 0)
+        {
+            final PreventionManager mgr = PreventionManager.getInstance();
+            final Prevention prevention = mgr.getPrevention(args[0]);
+            if (prevention != null)
+            {
+                mgr.disablePrevention(prevention);
+                prevention.reloadConfig();
+                mgr.enablePrevention(prevention);
+                sender.sendMessage(ChatColor.GREEN + "The prevention " + prevention.getName() + " was successfully reloaded!");
+            }
+            else
+            {
+                sender.sendMessage(ChatColor.RED + "The given preventions is not available!");
+            }
+        }
+        else
+        {
+            final Plugin plugin = getBase().getPlugin();
+            final PluginManager pm = plugin.getServer().getPluginManager();
+            pm.disablePlugin(plugin);
+            pm.enablePlugin(plugin);
 
-        sender.sendMessage(ChatColor.GREEN + plugin.getDescription().getName() + " successfully reloaded!");
+            sender.sendMessage(ChatColor.GREEN + plugin.getDescription().getName() + " successfully reloaded!");
+        }
 
         return true;
     }